Extract computeHypothesisConflict from state.tsx into Hypothesis.ts
Pure refactor; no user-visible behavior change. The hypothesis-conflict
banner aggregation that decides between `directly-contradicted` and
`jointly-conflicting` (and the `HypothesisConflict` /
`HypothesisConflictEntry` shape it produces) was inlined inside a
React `useMemo` in `state.tsx`. That made it impossible to unit-test
without mounting `<ClueProvider>`, which forced the existing per-branch
coverage to live in `state.test.tsx` as a heavy `renderHook` walk.

Move the function and its types to `src/logic/Hypothesis.ts` so the
aggregation can be tested as pure logic. `state.tsx` now imports
`computeHypothesisConflict` and re-exports `HypothesisConflict` so
existing consumers (`ContradictionBanner` imports the type from
`../state`) keep compiling without a path change.

Verified: `pnpm typecheck` clean, full `pnpm test` 1127/1127 green —
the React-layer behavior is unchanged.
